backend/requirements-deploy.txt
# ARK DevSecOps AI — deployment requirements
# This strips llama-cpp-python (needs C++ build) + test packages for cloud deploy

# ── Core Framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# ── Authentication / Security ─────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.28.1

# ── Git / Repository Cloning ──────────────────────────────────────────────────
gitpython==3.1.43

# ── AI Integration ────────────────────────────────────────────────────────────
google-generativeai==0.8.3
# llama-cpp-python is EXCLUDED — Mythos falls back to Heuristic tier on cloud
huggingface-hub==0.27.0

# ── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Redis (JWT denylist) ──────────────────────────────────────────────────────
redis==5.2.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
